requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
idna==2.8
lxml==4.3.4
Pillow==6.1.0
PixivPy==3.3.7
requests==2.22.0
soupsieve==1.9.2
urllib3==1.25.3
